guiTest/requirements.txt
pillow
tk
tkvideoPlayer
tkvideo
numpy
opencv-python